fixing compilation warnings

+++ src/buildmaster_utils.cc
@@ -60,7 +60,7 @@ NNPDF::dataInfoRaw readMeta(string setname)
 
 void Buildmaster::CommonData::SetData( unsigned int index, double datapoint)
 {
-    if (index >= fNData)
+    if (index >= static_cast<unsigned int>(fNData))
         throw NNPDF::RuntimeException("SetData", "Requested fill index: "+to_string(index) + " is out of bounds");
     if (fData[index] == fData[index]) // Data value is not a NaN (unset values are NaN)
         throw NNPDF::RuntimeException("SetData", "Requested fill index: "+to_string(index) + " is already filled");
@@ -69,7 +69,7 @@ void Buildmaster::CommonData::SetData( unsigned int index, double datapoint)
 
 void Buildmaster::CommonData::SetStatisticalError( unsigned int index, double staterror )
 {
-    if (index >= fNData)
+    if (index >= static_cast<unsigned int>(fNData))
         throw NNPDF::RuntimeException("SetStatisticalError", "Requested fill index: "+to_string(index) + " is out of bounds");
     if (fStat[index] == fStat[index]) // Value is not a NaN (unset values are NaN)
         throw NNPDF::RuntimeException("SetStatisticalError", "Requested fill index: "+to_string(index) + " is already filled");
